requirements.txt
flask
flask-cors>=3.0.0
gunicorn
google-generativeai
